Move function `read` into `textual_input_port` and remove `reader.hpp`
