+++ src/kernel/character.cpp
@@ -14,7 +14,6 @@
    limitations under the License.
 */
 
-#include <meevax/kernel/character.hpp>
 #include <meevax/kernel/eof.hpp>
 #include <meevax/kernel/error.hpp>
