+++ src/kernel/textual_input_port.cpp
@@ -16,9 +16,11 @@
 
 #include <meevax/kernel/environment.hpp>
 #include <meevax/kernel/eof.hpp>
+#include <meevax/kernel/homogeneous_vector.hpp>
 #include <meevax/kernel/interaction_environment.hpp>
 #include <meevax/kernel/string.hpp>
 #include <meevax/kernel/textual_input_port.hpp>
+#include <meevax/kernel/vector.hpp>
 
 namespace meevax
 {
@@ -50,6 +52,38 @@ inline namespace kernel
                                           '{',  // 0x7B
                                           '|',  // 0x7C
                                           '}'); // 0x7D
+  }
+
+  struct datum_label
+  {
+    std::string const n;
+
+    template <typename... Ts>
+    explicit datum_label(Ts&&... xs)
+      : n { std::forward<decltype(xs)>(xs)... }
+    {}
+  };
+
+  auto circulate(object const& xs, object const& x, std::string const& n) -> void
+  {
+    if (xs.is<pair>())
+    {
+      circulate(car(xs), x, n);
+
+      if (cdr(xs).is<datum_label>() and cdr(xs).as<datum_label>().n == n)
+      {
+        cdr(xs) = x;
+      }
+      else
+      {
+        circulate(cdr(xs), x, n);
+      }
+    }
+  }
+
+  auto circulate(object const& xs, std::string const& n) -> void
+  {
+    return circulate(xs, xs, n);
   }
 
   auto textual_input_port::get() -> object
@@ -119,7 +153,283 @@ inline namespace kernel
   {
     try
     {
-      return interaction_environment().as<environment>().read(*this);
+      auto & is = static_cast<std::istream &>(*this);
+
+      while (not character::is_eof(is.peek()))
+      {
+        switch (auto const c1 = is.peek())
+        {
+        case '\t': // 0x09
+        case '\n': // 0x0A
+        case '\v': // 0x0B
+        case '\f': // 0x0C
+        case '\r': // 0x0D
+        case ' ':  // 0x20
+          is.ignore(1);
+          break;
+
+        case '"':  // 0x22
+          return make(read_string_literal());
+
+        case '#':  // 0x23
+          switch (auto const c2 = is.ignore(1).peek())
+          {
+          case '!': // SRFI 22
+            is.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+            return read();
+
+          case ',': // SRFI 10
+            is.ignore(1);
+            return interaction_environment().as<environment>().evaluate(read());
+
+          case ';': // SRFI 62
+            is.ignore(1);
+            read(); // Discard an expression.
+            return read();
+
+          case '"':
+            return make_symbol(read_string_literal());
+
+          case '0':
+          case '1':
+          case '2':
+          case '3':
+          case '4':
+          case '5':
+          case '6':
+          case '7':
+          case '8':
+          case '9':
+            switch (auto n = take_digits(); is.peek())
+            {
+            case '#':
+              is.ignore(1);
+
+              if (auto iter = datum_labels.find(n); iter != std::end(datum_labels))
+              {
+                return iter->second;
+              }
+              else
+              {
+                throw read_error(make<string>("it is an error to attempt a forward reference"),
+                                 make<string>(lexical_cast<std::string>('#', n, '#')));
+              }
+
+            case '=':
+              is.ignore(1);
+
+              if (auto [iter, success] = datum_labels.emplace(n, make<datum_label>(n)); success)
+              {
+                if (let const& xs = read(); xs != iter->second)
+                {
+                  circulate(xs, n);
+                  datum_labels.erase(n);
+                  return xs;
+                }
+                else
+                {
+                  return unit;
+                }
+              }
+              else
+              {
+                throw read_error(make<string>("duplicated datum-label declaration"),
+                                 make<string>(n));
+              }
+
+            default:
+              throw read_error(make<string>("unknown discriminator"),
+                               make<string>(lexical_cast<std::string>('#', n, is.get())));
+            }
+
+          case 'b':
+            return make_number(is.ignore(1).peek() == '#' ? lexical_cast<std::string>(read()) : take_token(), 2);
+
+          case 'c': // Common Lisp
+            is.ignore(1);
+            return [](let const& xs)
+            {
+              return make<complex>(tail(xs, 0).is<pair>() ? xs[0] : e0,
+                                   tail(xs, 1).is<pair>() ? xs[1] : e0);
+            }(read());
+
+          case 'd':
+            return make_number(is.ignore(1).peek() == '#' ? lexical_cast<std::string>(read()) : take_token(), 10);
+
+          case 'e':
+            is.ignore(1);
+            return exact(read()); // NOTE: Same as #,(exact (read))
+
+          case 'f':
+            is.ignore(1);
+
+            switch (auto const digits = take_digits(); std::stoi(digits))
+            {
+            case 32:
+              return make<f32vector>(read());
+
+            case 64:
+              return make<f64vector>(read());
+
+            default:
+              take_token();
+              return f;
+            }
+
+          case 'i':
+            is.ignore(1);
+            return inexact(read()); // NOTE: Same as #,(inexact (read))
+
+          case 'o':
+            return make_number(is.ignore(1).peek() == '#' ? lexical_cast<std::string>(read()) : take_token(), 8);
+
+          case 's':
+            is.ignore(1);
+
+            switch (auto const digits = take_digits(); std::stoi(digits))
+            {
+            case 8:
+              return make<s8vector>(read());
+
+            case 16:
+              return make<s16vector>(read());
+
+            case 32:
+              return make<s32vector>(read());
+
+            case 64:
+              return make<s64vector>(read());
+
+            default:
+              throw read_error(make<string>("An unknown literal expression was encountered"),
+                               make<string>(lexical_cast<std::string>("#s", digits)));
+            }
+
+          case 't':
+            take_token();
+            return t;
+
+          case 'u':
+            is.ignore(1);
+
+            switch (auto const digits = take_digits(); std::stoi(digits))
+            {
+            case 8:
+              return make<u8vector>(read());
+
+            case 16:
+              return make<u16vector>(read());
+
+            case 32:
+              return make<u32vector>(read());
+
+            case 64:
+              return make<u64vector>(read());
+
+            default:
+              throw read_error(make<string>("An unknown literal expression was encountered"),
+                               make<string>(lexical_cast<std::string>("#u", digits)));
+            }
+
+          case 'x':
+            return make_number(is.ignore(1).peek() == '#' ? lexical_cast<std::string>(read()) : take_token(), 16);
+
+          case '(':
+            return make<vector>(read());
+
+          case '\\':
+            is.putback(c1);
+            return make(read_character_literal());
+
+          case '|': // SRFI 30
+            is.ignore(1);
+            take_nested_block_comment();
+            return read();
+
+          default:
+            throw read_error(make<string>("unknown discriminator"), make<character>(c2));
+          }
+
+        case '\'': // 0x27
+          is.ignore(1);
+          return list(make_symbol("quote"), read());
+
+        case '(':  // 0x28
+          try
+          {
+            is.ignore(1);
+
+            if (let const& x = read(); x.is<eof>())
+            {
+              return x;
+            }
+            else
+            {
+              is.putback('('); // modifying putback (https://en.cppreference.com/w/cpp/io/basic_istream/putback)
+              return cons(x, read());
+            }
+          }
+          catch (std::integral_constant<char, ')'> const&)
+          {
+            return unit;
+          }
+          catch (std::integral_constant<char, '.'> const&)
+          {
+            let const x = read();
+            is.ignore(std::numeric_limits<std::streamsize>::max(), ')');
+            return x;
+          }
+
+        case ')':  // 0x29
+          is.ignore(1);
+          throw std::integral_constant<char, ')'>();
+
+        case ',':  // 0x2C
+          switch (is.ignore(1); is.peek())
+          {
+          case '@':
+            is.ignore(1);
+            return list(make_symbol("unquote-splicing"), read());
+
+          default:
+            return list(make_symbol("unquote"), read());
+          }
+
+        case ';':  // 0x3B
+          is.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+          break;
+
+        case '`':  // 0x60
+          is.ignore(1);
+          return list(make_symbol("quasiquote"), read());
+
+        case '|':  // 0x7C
+          return make_symbol(read_string_literal());
+
+        case '[':  // 0x5B
+        case ']':  // 0x5D
+        case '{':  // 0x7B
+        case '}':  // 0x7D
+          throw read_error(make<string>("left and right square and curly brackets (braces) are reserved for possible future extensions to the language"),
+                           make<character>(c1));
+
+        default:
+          if (auto const& token = take_token(); token == ".")
+          {
+            throw std::integral_constant<char, '.'>();
+          }
+          else try
+          {
+            return make_number(token, 10);
+          }
+          catch (std::invalid_argument const&)
+          {
+            return make_symbol(token);
+          }
+        }
+      }
+
+      return eof_object;
     }
     catch (eof const&)
     {
